refactor(tests): changed the tests folder + currency formatter
